Exception handling
Added basic bad input/exception handling.

Checking for user input, redirect if response not recognized.
Checking if the list has an ID or not for the counter. For new files or if all tasks have been deleted.
Checking if the json file exists, create if doesnt.

Cleaning out old debugging code.